+++ lib/src/main/java/com/mammb/code/piecetable/buffer/ChannelBuffer.java
@@ -16,7 +16,7 @@
 package com.mammb.code.piecetable.buffer;
 
 import com.mammb.code.piecetable.array.ChannelArray;
-import com.mammb.code.piecetable.array.IntArray;
+import com.mammb.code.piecetable.array.LongArray;
 import java.io.Closeable;
 import java.io.IOException;
 import java.nio.channels.SeekableByteChannel;
@@ -41,7 +41,7 @@ public class ChannelBuffer implements Buffer, Closeable {
     private final short pilePitch;
 
     /** The piles. */
-    private final int[] piles;
+    private final long[] piles;
 
     /** The lru cache. */
     private final LruCache cache;
@@ -54,7 +54,7 @@ public class ChannelBuffer implements Buffer, Closeable {
      * @param pilePitch the size of pitch
      * @param piles the piles
      */
-    private ChannelBuffer(ChannelArray ch, int length, short pilePitch, int[] piles) {
+    private ChannelBuffer(ChannelArray ch, int length, short pilePitch, long[] piles) {
         this.ch = ch;
         this.length = length;
         this.pilePitch = pilePitch;
@@ -105,7 +105,7 @@ public class ChannelBuffer implements Buffer, Closeable {
     static Buffer of(SeekableByteChannel channel, short pitch, Consumer<byte[]> consumer) {
         var ch = ChannelArray.of(channel);
         var charCount = 0;
-        var piles = IntArray.of(ch.length() / pitch);
+        var piles = LongArray.of(Math.toIntExact(ch.length() / pitch));
         for (int i = 0; i < ch.length(); i++) {
             if (charCount++ % pitch == 0) {
                 piles.add(i);
@@ -126,20 +126,20 @@ public class ChannelBuffer implements Buffer, Closeable {
 
 
     @Override
-    public int length() {
+    public long length() {
         return length;
     }
 
 
     @Override
-    public byte[] charAt(int index) {
-        int rawIndex = asIndex(index);
+    public byte[] charAt(long index) {
+        long rawIndex = asIndex(index);
         return Utf8.asCharBytes(ch.get(rawIndex, Math.min(rawIndex + 4, ch.length())), 0);
     }
 
 
     @Override
-    public byte[] bytes(int rawStart, int rawEnd) {
+    public byte[] bytes(long rawStart, long rawEnd) {
         return ch.get(rawStart, rawEnd);
     }
 
@@ -151,13 +151,13 @@ public class ChannelBuffer implements Buffer, Closeable {
 
 
     @Override
-    public Buffer subBuffer(int start, int end) {
+    public Buffer subBuffer(long start, long end) {
         return ReadBuffer.of(ch.get(asIndex(start), asIndex(end)));
     }
 
 
     @Override
-    public int asIndex(int index) {
+    public long asIndex(long index) {
         if (index == length) {
             return ch.length();
         }
@@ -165,8 +165,8 @@ public class ChannelBuffer implements Buffer, Closeable {
         if (cached.isPresent()) {
             return cached.get();
         }
-        int i = piles[index / pilePitch];
-        int remaining = index % pilePitch;
+        long i = piles[Math.toIntExact(index / pilePitch)];
+        int remaining = Math.toIntExact(index % pilePitch);
         for (; remaining > 0 && i < ch.length(); remaining--, i++) {
             i += (Utf8.followsCount(ch.get(i)) - 1);
         }
